pyproject.toml
[tool.poetry]
name = "skyagent"
version = "0.1.0"
description = ""
authors = ["David Tokar <1413david@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
docstring-parser = "^0.16"
rich = "^13.9.4"
anthropic = "^0.43.0"
pexpect = "^4.9.0"
googlesearch-python = "^1.3.0"
html2text = "^2024.2.26"
identify = "^2.6.7"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
pandas = "^2.2.3"
mammoth = "^1.9.0"
pymupdf = "^1.25.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.25.3"
vcrpy = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
]